refactor(cli): extract auth-session manager into server-auth.ts
#354 pass 4. Lifts the local auth-session logic out of startServer into a
createAuthSession(cloudApiBaseUrl) factory: readLocalAuthSession, isAuthValid
(with its 5-min TTL cache), clearLocalAuthSession, and fetchCloudApiWithLocalAuth,
plus the internal getAuthCandidates / invalidateAuthCache / validatedAuth state.

startServer now destructures the four route-facing methods from the factory;
all call sites are unchanged. Dropped the now-unused loadAllAuthTokens /
loadAnyAuthToken / removeAuthToken imports from server.ts (they moved with the
helpers).

server.ts: 3219 -> 3099 lines; new server-auth.ts is 156. No behavior change —
340 CLI tests pass, typecheck + lint clean, server boot verified via the
editor-server + cli-smoke e2e.

Refs #354
